requirements.txt
Flask
loguru
requests
pillow
numpy